# backend/swagger.yaml
openapi: 3.0.3
info:
  title: Olympia HR Platform API
  description: |
    API REST pour la plateforme intelligente de gestion des ressources humaines.
    Conforme à la réglementation tunisienne du travail et de la paie.
    
    ## Fonctionnalités
    - Gestion employés avec validation Joi
    - Présence biométrique anti-fraude
    - Heures supplémentaires (125%, 150%, 200%)
    - Paie conforme IRPP 2025 + CNSS
    - Congés workflow complet
    - Déclarations légales (CNSS, IR)
    - Analyse sentiment IA
    
  version: 1.0.0
  contact:
    name: Support Olympia HR
    email: support@olympia-hr.tn
servers:
  - url: http://localhost:5000/api
    description: Serveur de développement
  - url: https://api.olympia-hr.tn/api
    description: Serveur de production
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT obtenu via /auth/login
...
paths:
...
  /legal/cnss/{month}:
    get:
      tags:
        - Legal
      summary: Bordereau CNSS mensuel
      description: Génère fichier Excel avec cotisations CNSS (employé 9.18% + employeur 16.57%)
      security:
        - bearerAuth: []
      parameters:
        - name: month
          in: path
          required: true
          schema:
            type: string
            pattern: '^\d{4}-\d{2}$'
          example: "2025-12"
      responses:
        '200':
          description: Fichier Excel généré
          content:
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
              schema:
                type: string
                format: binary
        '404':
          description: Aucune paie pour ce mois
  /legal/ir-annual/{year}:
    get:
      tags:
        - Legal
      summary: Déclaration IR annuelle
      description: Génère fichier Excel récapitulatif IRPP retenu à la source
      security:
        - bearerAuth: []
      parameters:
        - name: year
          in: path
          required: true
          schema:
            type: string
            pattern: '^\d{4}$'
          example: "2025"
      responses:
        '200':
          description: Fichier Excel généré
          content:
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
              schema:
                type: string
                format: binary
  /legal/work-certificate/{employee_id}:
    get:
      tags:
        - Legal
      summary: Attestation de travail
      description: Génère PDF officiel certifiant l'emploi
      security:
        - bearerAuth: []
      parameters:
        - name: employee_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: PDF généré
          content:
            application/pdf:
              schema:
                type: string
                format: binary
...